Updated all files to StormCustomMetricsRegistry.

+++ external/storm-metrics-micrometer/src/main/java/org/apache/storm/metric/IHistogram.java
@@ -19,7 +19,7 @@
 package org.apache.storm.metric;
 
 public interface IHistogram extends IStormMetric{
-    void update(int value);
+    void update(long value);
 
     long getCount();
 }
