+++ external/storm-metrics-micrometer/src/main/java/org/apache/storm/metric/dropwizard/StormDropwizardHistogram.java
@@ -30,7 +30,7 @@ public class StormDropwizardHistogram implements IHistogram {
     }
 
     @Override
-    public void update(int value) {
+    public void update(long value) {
         histogram.update(value);
     }
