+++ external/storm-metrics-micrometer/src/main/java/org/apache/storm/metric/micrometer/StormHistogram.java
@@ -30,7 +30,7 @@ public class StormHistogram implements IHistogram {
     }
 
     @Override
-    public void update(int value) {
+    public void update(long value) {
         summary.record(value);
     }
